% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{hbsir.add_attributes}
\alias{hbsir.add_attributes}
\title{Add attributes to HBSIR data}
\usage{
hbsir.add_attributes(df)
}
\arguments{
\item{df}{A data frame with 'ID' and 'year' columns.}
}
\value{
The data frame with added 'Province', 'County', 'Urban_Rural' columns.
}
\description{
This function adds province, county, and urban/rural attributes to the data frame based on ID and year.
}
