% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weight.R
\name{hbsir.add_weight}
\alias{hbsir.add_weight}
\title{Add household weights to a data frame}
\usage{
hbsir.add_weight(df)
}
\arguments{
\item{df}{A data frame that must include 'year' and 'ID' columns.}
}
\value{
The same data frame with an added 'Weight' column.
}
\description{
This function adds household sampling weights to a given data frame using either
household_information tables (if available) or an external fallback file.
}
