% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percentile_decile_quantile.R
\name{hbsir_add_percentile_decile_quantile}
\alias{hbsir_add_percentile_decile_quantile}
\title{Add Percentile, Decile, and Quantile to HBSIR Data}
\usage{
hbsir_add_percentile_decile_quantile(df, column)
}
\arguments{
\item{df}{A data frame containing the HBSIR data with a column to rank (e.g., "Gross_Income" or "Gross_Expenditure").}

\item{column}{The name of the column to calculate ranks on (e.g., "Gross_Income").}
}
\value{
A data frame with added columns: \verb{<column>_Percentile}, \verb{<column>_Decile}, and \verb{<column>_Quartile}.
}
\description{
This function adds percentile, decile, and quantile (using quartiles as an example) calculations
to a data frame containing HBSIR data, based on a specified income or expenditure column.
}
