% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_functions.R
\name{load.income}
\alias{load.income}
\title{Load income data for a given year}
\usage{
load.income(year, combine = TRUE, summarise = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{year}{The year of the data.}

\item{combine}{Logical. If TRUE, combine all tables.}

\item{summarise}{Logical. If TRUE, summarize the combined data.}

\item{verbose}{Logical. If TRUE, print messages.}

\item{...}{Additional arguments passed to load_table.}
}
\value{
A data.frame or list of data.frames, depending on parameters.
}
\description{
This function loads all cached income tables for a given year, with options to combine and summarize.
}
