% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_functions.R
\name{preload_income_available_tables}
\alias{preload_income_available_tables}
\title{Preload available income tables for a given year}
\usage{
preload_income_available_tables(year, ...)
}
\arguments{
\item{year}{The year of the data.}

\item{...}{Additional arguments passed to load_table.}
}
\value{
Invisible NULL; side effect is caching tables.
}
\description{
This function preloads missing tables for income data by checking the cache and downloading if necessary.
}
